Corrected latin translation of "stop"
Sistite is correct when addressing a group. Siste is correct for an individual. Unsure of which one to use I think Sistite makes more sense. That combined with the fact that hardly anyone knows it's here at all makes me not worry much about it. 
But still, Caesum is a pause, closer related to a "full stop" in the U.K. or a comma, not the same as "stop."