Fixed substitution (s///).
There were two problems with minification:

First, the substitution's matcher function (passed into `bot.eval`)
referenced an argument `eval` brought it, but which was not actually in
the function's lexical scope.
Fix is to explicitly use `arguments`.

Second, to return the result, the matcher simply left a variable
dangling around at the end of the function (as `eval` picks up the last
expression's value as the result), which was minified away.
Fix is to wrap it around in an IIFE so it won't appear to be a no-op.